поправила тесты на сторы и логику
