+++ .eslintrc
@@ -4,7 +4,8 @@
   "env": {
     "browser": true,
     "es6": true,
-    "node": true
+    "node": true,
+    "jest": true
   },
   "plugins": [
     "import",
