+++ postcss.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  // Add your installed PostCSS plugins here:
+  plugins: [
+    require('autoprefixer'),
+    require('postcss-flexbugs-fixes'),
+  ],
+};
