--- src/App/AppModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppModal';
