+++ src/App/PhoneInput/PhoneInput.tsx
@@ -30,8 +30,6 @@ const PhoneInput: React.FC<Props> = ({ masks, value, onChange }: Props) => {
     });
   }, [store.selectedCountryData]);
 
-  console.log({ value: value, refactor: storeNumber?.fields });
-
   const [openedPopup, setOpenedPopup] = React.useState(false);
 
   const handleCLosePopup = React.useCallback(() => {
