+++ src/shared/entities/phone/index.ts
@@ -1,2 +1,3 @@
 export * from './server';
+export * from './client';
 export * from './plugs';
