+++ src/shared/utils/keyCodes.ts
@@ -1,8 +1,6 @@
-const keyCodes = {
+export const keyCodes = {
   ENTER_KEYS: ['13', 'Enter'],
   ARROW_RIGHT_KEYS: ['26', 'ArrowRight'],
   ARROW_LEFT_KEYS: ['27', 'ArrowLeft'],
   BACKSPACE_KEYS: ['Backspace', 'Delete'],
 };
-
-export default keyCodes;
